don't implicity cast

+++ board/usb_comms.h
@@ -275,7 +275,7 @@ int usb_cb_control_msg(USB_Setup_TypeDef *setup, uint8_t *resp) {
 
     // **** 0xdc: set safety mode
     case 0xdc:
-      set_safety_mode(last_safety_mode, setup->b.wValue.w | (setup->b.wIndex.w << 16));
+      set_safety_mode(last_safety_mode, (uint32_t)setup->b.wValue.w | ((uint32_t)setup->b.wIndex.w << 16));
       break;
     case 0xfc:
       last_safety_mode = setup->b.wValue.w;
